Move transformers to their own folder
